# public/openapi.yaml
openapi: 3.0.3
info:
  title: Linkage Center API
  version: 1.1.0
  description: |
    เอกสาร API สำหรับระบบ Linkage Center และบริการตรวจสอบบัตร/สถานะประชาชน
    โดย Swagger UI ถูกตั้งค่าให้ใช้งานกับ Express 5 และไฟล์ openapi.yaml โดยตรง

    การยืนยันตัวตนของกลุ่ม Linkage รองรับ 2 รูปแบบ
    1. Authorization: Bearer <token>
    2. header ชื่อ token: <token>

    เพื่อความสะดวกในการทดสอบผ่าน Swagger UI แนะนำให้ใช้ BearerAuth เป็นหลัก
servers:
  - url: /
    description: Current server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ใช้กับ endpoint กลุ่ม Linkage ที่ต้องมี token หลัง login
    TokenHeader:
      type: apiKey
      in: header
      name: token
      description: รูปแบบเดิมของระบบ รองรับควบคู่กับ BearerAuth
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: body not variable is required
    ErrorEnvelope:
      type: object
      required: [errorNumber, errorMessage, data]
      properties:
        errorNumber:
          type: integer
          example: 401
        errorMessage:
          type: string
          nullable: true
          example: Unauthorized
        data:
          nullable: true
          oneOf:
            - type: object
            - type: 'null'
    StandardApiEnvelope:
      type: object
      required: [errorNumber, errorMessage, data]
      properties:
        errorNumber:
          type: integer
          example: 0
        errorMessage:
          type: string
          nullable: true
          example: null
        data:
          description: ข้อมูลที่ส่งกลับจากบริการต้นทาง
          oneOf:
            - type: object
            - type: array
              items: {}
            - type: string
            - type: integer
            - type: boolean
            - type: 'null'
...
    CheckDeathRequest:
      type: object
      required: [targetId, dob]
      properties:
        targetId:
          type: string
          pattern: '^\d{13}$'
          description: เลขประจำตัวประชาชน 13 หลัก
          example: '1234567890123'
        dob:
          type: string
          description: วันที่เกิดในรูปแบบที่ web service ต้องการ เช่น YYYYMMDD
          example: '19800101'
    CheckDeathResponseData:
      type: object
      properties:
        stCode:
          type: string
          nullable: true
          example: '0'
        stDesc:
          type: string
          example: พบข้อมูลบุคคล
        isError:
          type: boolean
          example: false
        isErrorMessage:
          type: string
          example: ''
    CheckDeathEnvelope:
      allOf:
        - $ref: '#/components/schemas/StandardApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CheckDeathResponseData'
    CheckLaserRequest:
      type: object
      required: [pid, fname, lname, dob, laser]
      properties:
        pid:
          type: string
          pattern: '^\d{13}$'
          description: เลขประจำตัวประชาชน 13 หลัก
          example: '1234567890123'
        fname:
          type: string
          description: ชื่อจริง
          example: สมชาย
        lname:
          type: string
          description: นามสกุล
          example: ใจดี
        dob:
          type: string
          description: วันเดือนปีเกิดในรูปแบบที่ service ต้องการ
          example: '19800101'
        laser:
          type: string
          description: เลข Laser ด้านหลังบัตรประชาชน
          example: JT0-1234567-89
    CheckLaserResponse:
      type: object
      properties:
        isError:
          type: boolean
          example: false
        isErrorDesc:
          type: string
          example: ''
        code:
          type: integer
          example: 0
        desc:
          type: string
          example: ตรวจสอบข้อมูลสำเร็จ
    CheckChipRequest:
      type: object
      required: [pid, chipno, bp1no]
      properties:
        pid:
          type: string
          pattern: '^\d{13}$'
          description: เลขประจำตัวประชาชน 13 หลัก
          example: '1234567890123'
        chipno:
          type: string
          description: หมายเลขชิปการ์ด (ChipNo)
          example: A1B2C3D4E5F6
        bp1no:
          type: string
          description: รหัส BP1 ตามที่ต้นทางกำหนดใช้
          example: '0101'
    CheckChipResponse:
      type: object
      properties:
        isError:
          type: boolean
          example: false
        isErrorDesc:
          type: string
          example: ''
        code:
          type: integer
          example: 0
        desc:
          type: string
          example: ตรวจสอบข้อมูลสำเร็จ
  responses:
    BadRequestError:
      description: คำขอไม่ถูกต้องหรือข้อมูลไม่ครบ
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ErrorEnvelope'
    InternalServerError:
      description: เกิดข้อผิดพลาดภายในระบบ
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ErrorEnvelope'
paths:
...
  /card/api/death:
    post:
      tags: [CheckPop]
      summary: ตรวจสอบสถานะบุคคล/บัตร (CheckDeathStatus)
      description: เรียก SOAP service เพื่อตรวจสอบสถานะจากเลขประชาชนและวันเกิด
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckDeathRequest'
            examples:
              sample:
                value:
                  targetId: '1234567890123'
                  dob: '19800101'
      responses:
        '200':
          description: ตรวจสอบสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckDeathEnvelope'
              examples:
                success:
                  value:
                    errorNumber: 0
                    errorMessage: null
                    data:
                      stCode: '0'
                      stDesc: พบข้อมูลบุคคล
                      isError: false
                      isErrorMessage: ''
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /card/api/checklaser:
    post:
      tags: [CheckLaser]
      summary: ตรวจสอบบัตรด้วยเลข Laser (CheckCardByLaser)
      description: ตรวจสอบข้อมูลบัตรจาก pid ชื่อ นามสกุล วันเกิด และเลข Laser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckLaserRequest'
            examples:
              sample:
                value:
                  pid: '1234567890123'
                  fname: สมชาย
                  lname: ใจดี
                  dob: '19800101'
                  laser: JT0-1234567-89
      responses:
        '200':
          description: ตรวจสอบสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckLaserResponse'
              examples:
                success:
                  value:
                    isError: false
                    isErrorDesc: ''
                    code: 0
                    desc: ตรวจสอบข้อมูลสำเร็จ
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /card/api/checkchip:
    post:
      tags: [CheckChip]
      summary: ตรวจสอบบัตรด้วย ChipNo (CheckCardByCID)
      description: ตรวจสอบข้อมูลบัตรจาก pid, chipno และ bp1no ผ่าน SOAP service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckChipRequest'
            examples:
              sample:
                value:
                  pid: '1234567890123'
                  chipno: A1B2C3D4E5F6
                  bp1no: '0101'
      responses:
        '200':
          description: ตรวจสอบสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckChipResponse'
              examples:
                success:
                  value:
                    isError: false
                    isErrorDesc: ''
                    code: 0
                    desc: ตรวจสอบข้อมูลสำเร็จ
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'